Add files via upload
Added a bunch of sections, learning how to use display flex and display grid and using grid-template-columns, which I havent seen or used before. Also using a different method in JS to toggle an open class on the body of the document to make the navigation slide onto the screen and off the screen. As well as change the navigation button from the 3 bars to an X. This project is teaching me a lot of more advanced CSS using the grid layout css rules