Add files via upload
Finished the main webpage, not responsive yet. Implemented some scripts from CDNJS for JQuery and Slick Carousel. Finished up the book now section, customer carousel and footer. I started to about CSS flex and grid in this project and it has helped further my understanding of layout of a webpage. I'm going to look into flex and grid in further depth to gain a better understanding off webpage layout as this is a powerful way of designing your site. I also learned a little bit about JS with the navigation button, applying a class to the whole body but have it only affect certain things that you've specified have that class in your CSS file. I also learned about scroll behaviour smooth which is a nice feature, that you don't need to write out in JS.